Implemented Google Cloud security rule plugin.
The implementation now supports:
 - Insert
 - Get
 - List
 - Delete

Although, this only applies to the Google Cloud firewall's security rule, an alternative to the public ip's security rule has not been found yet. Moreover, the firewall only supports the ether type IPv4, there are workarounds to support IPv6, but they are not trivial and would intensely deviate from the fogbow original design.

Therefore, an SecurityGroup abstraction was avoided, since (so far) there is not a well stablished relationship between the firewall and both fogbow's public ip and its security rule abstractions.